Squashed a hard bug

+++ include/ndpip/socket.h
@@ -20,8 +20,8 @@ int ndpip_setsockopt(int sockfd, int level, int optname, const void *optval, soc
 int ndpip_getsockopt(int sockfd, int level, int optname, const void *optval, socklen_t optlen);
 int ndpip_recv(int sockfd, struct ndpip_pbuf **pb, uint16_t count);
 int ndpip_send(int sockfd, struct ndpip_pbuf **pb, uint16_t count);
-int ndpip_free(int sockfd, struct ndpip_pbuf **pb, uint16_t len);
-int ndpip_alloc(int sockfd, struct ndpip_pbuf **pb, uint16_t len);
+int ndpip_free(int sockfd, struct ndpip_pbuf **pb, size_t len);
+int ndpip_alloc(int sockfd, struct ndpip_pbuf **pb, size_t len);
 #ifdef NDPIP_GRANTS_ENABLE
 int ndpip_cost(int sockfd, struct ndpip_pbuf **pb, uint16_t len, uint16_t *pb_cost);
 int ndpip_grants_get(int sockfd, uint32_t grants);
